requirements.txt
streamlit
streamlit-lottie
deep-translator
langdetect
emoji
pandas
matplotlib
seaborn
python-dotenv
openai
google-generativeai
youtube-comment-downloader
yt-dlp
requests
